Refactor Checkup and Consult pages for a better user experience
Improve the Checkup and Consult pages by refactoring the code and
updating the UI components. The changes include replacing the booking
image with a self image, adding a termo image, updating the AppHeading
component, and adjusting the layout and styling of various elements.
This refactor enhances the user experience and promotes proactive
self-health checkups.

Issue #42, #53